[REFACTOR] Pulling up outline to IExperienceManager
